pyproject.toml
[project]
name = "fahrgemeinschaften"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "flask>=3.1.2",
    "geopy>=2.4.1",
    "openrouteservice>=2.3.3",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "staticmap>=0.5.7",
]
